+++ kvm-ioctls/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming Release
 
+### Added
+
+- Added support for `KVM_HAS_DEVICE_ATTR` and `KVM_SET_DEVICE_ATTR` vm ioctl
+  on aarch64.
 - Plumb through KVM_CAP_X2APIC_API as X2ApicApi cap.
 
 ## v0.23.0